test: Move AI features tests to separate folder
