+++ packages/testing/playwright/tests/ui/ai/rag-callout.spec.ts
@@ -1,4 +1,4 @@
-import { test, expect } from '../../fixtures/base';
+import { test, expect } from '../../../fixtures/base';
 
 test.describe('RAG callout experiment', () => {
 	test.describe('NDV callout', () => {
